Move `create-store.js` and `url-handling.js` to a new directory `app-logic` (#1116)
