+++ src/app-logic/README.md
@@ -1,0 +1,9 @@
+# Application logic - /src/app-logic
+
+Most of the application logic is contained in reducers (`reducers/app.js` or
+`reducers/profile-view.js` usually).
+
+Therefore this folder is for when we need to extract some business logic, either
+because it's complicated or because it needs to be shared across the
+application.
+
