+++ src/app-logic/create-store.js
@@ -7,8 +7,8 @@
 import { createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 import { createLogger } from 'redux-logger';
-import reducers from './reducers';
-import type { Store } from './types/store';
+import reducers from '../reducers';
+import type { Store } from '../types/store';
 
 /**
  * Isolate the store creation into a function, so that it can be used outside of the
